Make it so that redis storage has the least s3 logic as possible

+++ piggy_store/storage/users/redis_storage.py
@@ -1,17 +1,11 @@
+from piggy_store.storage.users import User
+from piggy_store.exceptions import UserExistsError, UserDoesNotExistError
+from piggy_store.storage import EasyStorage
+
 import redis
 
-from piggy_store.exceptions import UserExistsError, UserDoesNotExistError
-from piggy_store.storage.users.user_entity import User
-from piggy_store.storage.users.storage import Storage as BaseStorage
-from piggy_store.storage.files import (
-    access_admin_storage,
-    access_user_storage,
-    compose_challenge_file_filename,
-    parse_challenge_file_filename
-)
 
-
-class Storage(BaseStorage):
+class Storage():
     __instance = None
 
     def __new__(cls, options, **kwargs):
@@ -27,7 +21,7 @@ class Storage(BaseStorage):
         return cls.__instance
 
     def __init__(self, *args, **kwargs):
-        pass
+        self.es = EasyStorage()
 
     def add_user(self, user):
         if self.conn.exists(user.username):
@@ -54,12 +48,9 @@ class Storage(BaseStorage):
             user = User(username, data['challenge'], data['answer'])
         else:
             # Do we have the user data at all?
-            file_storage = access_admin_storage()
-            challenge_file = file_storage.get_first_matching_file(compose_challenge_file_filename(username, ''))
-            if challenge_file:
-                challenge = file_storage.get_file_content(challenge_file).decode('utf-8')
-                answer = parse_challenge_file_filename(challenge_file.get_filename())['answer']
-                user = User(username, challenge, answer)
+            user = self.es.find_user_by_username(username)
+            if user:
+                # found, let's cache it
                 self.add_user(user)
 
         if user is None:
@@ -69,12 +60,6 @@ class Storage(BaseStorage):
 
     def remove_user_by_username(self, username):
         user = self.find_user_by_username(username)
-        user_file_storage = access_user_storage(user.username)
-        user_file_storage.remove_multiple((user_file_storage.get_files_list()))
-
-        admin_file_storage = access_admin_storage()
-        challenge_file_filename = compose_challenge_file_filename(user.username, user.answer)
-        f = admin_file_storage.build_file(challenge_file_filename)
-        admin_file_storage.remove_file(f)
-
-        self.delete_user(user)
+        if user:
+            self.es.remove_user(user)
+            self.delete_user(user)
